Move all CoreRegisterLinker together into SharingReceiver.
