+++ oap-server/server-starter/pom.xml
@@ -131,6 +131,11 @@
             <artifactId>storage-zipkin-plugin</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.skywalking</groupId>
+            <artifactId>jaeger-receiver-plugin</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <!-- storage module -->
 
         <!-- queryBuild module -->
